requirements.txt
Django==3.2.19
djangorestframework==3.14.0
psycopg2-binary==2.9.6
python-decouple==3.8
django-extensions==3.2.3